// zombiesoup-chesswin-front-e77511bd6471/src/app/common/constants/game-action-types.ts
export enum gameActionTypes {
	CONNECTION_ID = 'CONNECTION_ID',
	CONNECTION_ON = 'CONNECTION_ON',
	CONNECTION_OFF = 'CONNECTION_OFF',
	GAME_REQUEST_SEND = 'GAME_REQUEST_SEND',
	NEW_GAME_REQUEST = 'NEW_GAME_REQUEST',
	GAME_REQUEST_DECLINED = 'GAME_REQUEST_DECLINED',
	GAME_REQUEST_ACCEPTED = 'GAME_REQUEST_ACCEPTED',
	GAME_SUBSCRIBED = 'GAME_SUBSCRIBED',
	CLOSE_PRE_GAME_POPUP = 'CLOSE_PRE_GAME_POPUP',
	CHAT_MESSAGE = 'CHAT_MESSAGE',
	GAME_MOVE = 'GAME_MOVE',
	GAME_MOVE_FEN_ONLY = 'GAME_MOVE_FEN_ONLY',
	GAME_ABORT = 'GAME_ABORT',
	GAME_ABORTED_POPUP = 'GAME_ABORTED_POPUP',
	DRAW_REQUEST = 'DRAW_REQUEST',
	DRAW_REQUEST_DECLINE = 'DRAW_REQUEST_DECLINE',
	DRAW_REQUEST_ACCEPT = 'DRAW_REQUEST_ACCEPT',
	DRAW_REQUEST_SENT = 'DRAW_REQUEST_SENT',
	GAME_ENDED = 'GAME_ENDED',
	FINISH_GAME = 'FINISH_GAME'
}